test: consolidate OpenCode report fallback integration
